Move update logic to UpdateChecker
