+++ MudaeFarm/UpdateChecker.cs
@@ -119,5 +119,52 @@ namespace MudaeFarm
                     await entryStream.CopyToAsync(stream);
             }
         }
+
+        public static async Task InstallUpdateAsync(string path)
+        {
+            Log.Info($"Upgrading to v{CurrentVersion.ToString(3)}...");
+
+            // wait for old process to exit
+            await Task.Delay(TimeSpan.FromSeconds(1));
+
+            var dir = Directory.CreateDirectory(path);
+
+            try
+            {
+                // delete all old files
+                foreach (var file in dir.EnumerateFiles())
+                    file.Delete();
+            }
+            catch
+            {
+                // ignored
+            }
+
+            // copy ourselves in
+            foreach (var file in new DirectoryInfo(AppDomain.CurrentDomain.BaseDirectory).EnumerateFiles())
+            {
+                try
+                {
+                    var dest = Path.Combine(dir.FullName, file.Name);
+
+                    file.CopyTo(dest, true);
+
+                    Log.Debug(dest);
+                }
+                catch (Exception e)
+                {
+                    Log.Error($"Could not copy: {file.Name}", e);
+                }
+            }
+
+            // run new installation
+            Process.Start(new ProcessStartInfo
+            {
+                FileName        = dir.EnumerateFiles("*.exe").First().FullName,
+                UseShellExecute = false
+            });
+
+            Process.GetCurrentProcess().Kill();
+        }
     }
 }
